Remade card components with single standard card component (#345)
* remade card components with standardcard component

* lint rgba colors

* refactored cop cards into the landing page

* renamed standard card to protocard

* lint fixes

* lint fixes

* lint fixes

* lint fixes

* corrected export for cards scss

* renamed protocard to standard card
