+++ app/frontend/src/components/Cards/_StandardCard.scss
@@ -1,0 +1,6 @@
+.card {
+  border-radius: 16px;
+  box-shadow: 0 2px 8px rgb(0 0 0 / 25%);
+  box-sizing: border-box;
+  padding: 40px;
+}
