+++ app/frontend/src/components/Cards/_index.scss
@@ -1,2 +1,2 @@
-@forward "Cards";
+@forward "StandardCard";
 @forward "CircleCard";
