+++ app/frontend/src/components/Inputs/_Calendar.scss
@@ -49,7 +49,7 @@ $cell-border: 2px;
     @include calendar-header-row;
   }
 
-  & > div:not(:first-child):not(:last-child) {
+  & > div:not(:first-child, :last-child) {
     height: 2 * ($cell-height + 2 * $cell-border);
   }
 }
@@ -102,7 +102,7 @@ $cell-border: 2px;
   }
 
   &:hover:not(.selected) {
-    background-color: rgba($color-green, 0.5);
+    background-color: rgba($color-green, 50%);
   }
 
   & > * {
