+++ app/frontend/src/components/Inputs/_Dropdown.scss
@@ -35,6 +35,6 @@
   }
 
   &:hover {
-    background-color: rgba($color-blue-dark, 0.15);
+    background-color: rgba($color-blue-dark, 15%);
   }
 }
