+++ app/frontend/src/pages/LandingPage/LandingPageCop.tsx
@@ -2,14 +2,8 @@
 import React, { useState, useEffect } from "react";
 
 // Internal Imports
-import {
-  Button,
-  CircleCard,
-  CopCard,
-  Dialog,
-  InnerCopCard,
-  InnerCopNavCard,
-} from "components/components";
+import { Button, CircleCard, Dialog } from "components/components";
+import { CopCard, InnerCopCard, InnerCopNavCard } from "./LandingPageCopCards";
 import { copDatum, fetchAllCopData, fetchCopDataById } from "./copData";
 
 function LandingPageCop() {
