+++ app/frontend/src/pages/LandingPage/_index.scss
@@ -1,2 +1,3 @@
 @forward "LandingPageIntro";
 @forward "LandingPageCop";
+@forward "LandingPageCopCards";
